requirements.txt
Django==4.2
Pillow
geojson
django-admin-sortable2